fix msg error

+++ ecomap/www/admin_views.py
@@ -896,7 +896,8 @@ def delete_problem_type():
         file_name = db.get_problem_type_picture(data['problem_type_id'])
         static_url = '/media/image/markers'
         f_path = os.environ['STATICROOT'] + static_url
-        os.remove(os.path.join(f_path, file_name[0]))
+        if os.path.exists(os.path.join(f_path, file_name[0])):
+            os.remove(os.path.join(f_path, file_name[0]))
         db.delete_problem_type(data['problem_type_id'])
         if not db.get_problem_type_by_id(data['problem_type_id']):
             response = jsonify(msg='Success'), 200
@@ -947,7 +948,7 @@ def add_problem_type():
                                     problem_type_radius)
                 response = jsonify(msg='Success'), 200
             else:
-                response = jsonify(msg='Wrong data'), 400
+                response = jsonify(msg='Incorrect photo'), 400
     else:
         response = jsonify(msg='Incorrect data'), 400
     # if valid['status']:
@@ -999,7 +1000,8 @@ def edit_problem_type():
             img_file.save(os.path.join(f_path, file_name))
             old_file_name = db.get_problem_type_picture(
                 data['problem_type_id'])
-            os.remove(os.path.join(f_path, old_file_name[0]))
+            if os.path.exists(os.path.join(f_path, old_file_name[0])):
+                os.remove(os.path.join(f_path, old_file_name[0]))
             problem_type_id = request.form['problem_type_id']
             problem_type_name = request.form['problem_type_name']
             problem_type_radius = request.form['problem_type_radius']
@@ -1007,7 +1009,7 @@ def edit_problem_type():
                                    problem_type_name, problem_type_radius)
             response = jsonify(msg='Success'), 200
         else:
-            response = jsonify(msg='Wrong data'), 400
+            response = jsonify(msg='Incorrect photo'), 400
     else:
         response = jsonify(msg='Incorrect data'), 400
     # data = secure_filename(img_file.filename)
